feat(npm): use "ci" (cleaninstall) if lockfiles are present (#580)
Lockfiles, explicitly designed for reproducible builds, are ignored by
the npm packager implementation of the plugin. Thus `sls package` will
attempt to install different versions of dependencies after resolving
them if it can (due to `npm install`).

This makes it harder to achieve reproducible builds and reliable
deployments. When a lockfile is present for npm, use the cleaninstall
command instead to install from the lockfile (also saving time on
resolution). Use the pattern in the Yarn packager implementation,
allowing this behavior to be tweaked by the use of packager options
`ignoreLockfile`. Since that configuration defaults to `false` [1], this
is a breaking change for npm users.

[1]: https://www.serverless.com/plugins/serverless-esbuild#packager-options